// src/types/global.ts
export interface IProduct {
  id: number;
  title: string;
  price: number;
  description: string;
  category: string;
  image: string;
  rating: {
    rate: number;
    count: number;
  };
  quantity: number;
}

export interface IGallery {
  src: string;
  alt: string;
}

export type SortOptions = "desc" | "asc" | "cheapest" | "expensive";